+++ .gitignore
@@ -1,4 +1,5 @@
 
+.idea
 node_modules
 # OSX
 .DS_STORE
