+++ v1.1_microservice/app.py
@@ -1,12 +1,11 @@
 import os
 import flask
 from flask import Flask, request
-from werkzeug.middleware.proxy_fix import ProxyFix
+# from werkzeug.middleware.proxy_fix import ProxyFix
 import json
 import time
 from datetime import datetime
 import sys
-sys.path.append('src')
 from DocuDive import loadModel, getResult
 
 
@@ -25,7 +24,7 @@ def predict(body: dict, llm):
 
     return result
 
-app = Flask(__name__)
+app = Flask(__name__, static_url_path='', static_folder='web/static')
 
 # Load the model by reading the `SM_MODEL_DIR` environment variable
 # which is passed to the container by SageMaker (usually /opt/ml/model).
@@ -33,25 +32,27 @@ llm = model_fn()
 
 # Since the web application runs behind a proxy (nginx), we need to
 # add this setting to our app.
-app.wsgi_app = ProxyFix(
-    app.wsgi_app, x_for=1, x_proto=1, x_host=1, x_prefix=1
-)
+# app.wsgi_app = ProxyFix(
+#     app.wsgi_app, x_for=1, x_proto=1, x_host=1, x_prefix=1
+# )
 
 
-@app.route("/ping", methods=["GET"])
-def ping():
+@app.route("/keepalive")
+def keepalive():
     """
     Healthcheck function.
     """
-    return "pong"
+    return "up"
 
 
-@app.route("/invocations", methods=["POST"])
+@app.route("/api/ask", methods=["POST"])
 def invocations():
     """
     Function which responds to the invocations requests.
     """
     body = request.json
+    if not body.query:
+        return "Missing 'query' in POST JSON body"
     isSuccess = True
     result = {}
     try:
@@ -73,4 +74,5 @@ def invocations():
     return flask.Response(response=responsejson, status=200, mimetype='application/json')
 
 if __name__ == "__main__":
-    app.run(host='localhost',port=8080)
+    # Using host="localhost" will not connect to the public IP in AWS EC2. You have to use 0.0.0.0
+    app.run(host='0.0.0.0',port=8080)
